added SDL2 include paths and fixed a few compiler warnings
changed type from uint32 to uintptr_t when passing a pointer as an
int to an array.

+++ runtime/info/src/outputredir.h
@@ -136,7 +136,7 @@ class COutputRedir : public ILTOutputRedir
 		uint32		m_PrintColor;
 
 		// log file handles
-		uint32		m_pLogFiles[MAX_LOG_FILES];
+		uintptr_t   m_pLogFiles[MAX_LOG_FILES];
 
 		// nesting levels
 		uint8		m_NestLevels[OUTPUT_REDIR_ALL];
